+++ src/system/actor_error.rs
@@ -10,4 +10,11 @@ pub enum ActorError {
     #[error("Actor exists, but is not of the expected Type")]
     InvalidActorTypeError,
 
+    /// Triggered by [ActorBuilder.spawn](../prelude/struct.ActorBuilder.html#method.spawn) if the actor can't be spawned, because the thread-pool already reached the configured actor limit
+    #[error("Actor could not be started, because thread-pool is full")]
+    ThreadPoolHasTooManyActorsError,
+
+    /// Triggered by [ActorBuilder.spawn](../prelude/struct.ActorBuilder.html#method.spawn) if the actor can't be spawned, because the thread-pool does not exist
+    #[error("Actor could not be started, because thread-pool does not exist")]
+    ThreadPoolDoesNotExistError,
 }